[#312] Refactor anonymous functions, use print()
fixes #312
